Ui locale tests (#38)
When investigating #37, I noticed that the i18n tests were perhaps lacking in what they tested. They were only testing that Markdown was being processed to HTML, but not checking WHAT what was being output. I've written a new series of tests that does THIS:

* Loops through all files in the locales directory
* Finds the locale and values from the YAML in each file
* Sets the I18n gem to the locale
* Processes the hints and save Markdown files for each locale
* Checks that the source strings in the YAML are present in the processed HTML

This does NOT fix the original issue. This is intended to be an enhancement to the test suite which might help us spot i18n failures quicker in future.